Add snapshot of EA Versioning / Release file
This document has been maintained in real time on another machine, so
here's a snapshot with today's date (now that this file is no longer in
the line of production this is unlikely to change). This tracks all the
changes since the time the EA file was elevated to what was then called
"Beta" (now "Substantive") for certain sections.