Completed draft of FIBO-IND Specification
The completed written specification document, hopefully ready to go.

I made a decision about the "Changes to Adopted Specifications: I
described in Chapter 3 the exact changes to make within the modeling
tool, but then added an Annex to render the actual changes you would see
in FIBO Foundations - this is in order to preserve the headings and page
format appearances. It is Informative while Chapter 3 is Normative.