Addition of UML XMI Machine Readables
UML XMI for Indices and Indicators, and for Foundations changes in
Indices and Indicators submission. This is the folder tracking the
precise deliveries made to OMG, and is NOT a place where things are
controlled or changed from. Since it tracks the IND submission this
directory also include Foundations material which is in support of that
submission - it does not reflect the namespace structure itself, which
is reflected elsewhere in this folder structure.