VOM Project Files Structure
Created a clone of the working VOM Project Folders structure for FIBO as
it currently stands on MB machine.

also included the techprocess structure - probably do not have the
latest dated SpecificationMetadata but this is what is on MB machine
(one dated and one undated variant). This may differ from what is on
Elisa machine and this may need to be dealt with in this Repository
record of what should be the workable, cloneable VOM project set.

The above needs to be discussed as part of the Pull Request processing
for this change.