Two revised About files for FIBO BE Specification (AboutBE and the version-specific AboutBE-1.0).
EDM Council JIRA Issue BE-1 cover this and the previously uploaded BE Specification and BE Module About files.

These have been placed directly in yellow for up to date access for all. A separate branch for the resolution and testing of EDMC JIRA BE-1 has not been created at this time.