BE-23, FBEF-31 - Testing with various reasoners and test data shows that reasoning over the hasRole relationship in some circumstances returns incorrect results. Replacing the hasRole property with isPlayedBy corrects this issue.
Additionally, in FND, the Control class changed names to ControllingCapacity, and the corresponding references have been updated in this ontology to reflect the change. Similarly, the domain of the holds property in FND/Relations was modified to be an AutonomousAgent, so properties of relative parties in ControlParties that were subproperties of holds were modified to remove this subproperty relationship.

Also, a Percentage class was added to FND/BusinessFacingTypes as a part of the FND revision, which is a better fit than the percentageValue datatype as the basis for representing the DegreeOfControl an EntityControllingParty has over another organization. The data property has been eliminated in favor of an object property between DegreeOfControl and Percentage, called hasPercentageControl, and a restriction on that property.