This issue affects section 9.6.4, Corporate Control Ontology, in the FIBO BE 1.0 Alpha specification.
Two occurrences of hasRole have been replaced with isPlayedBy, one in the definition of ControlledCompany and the other in the extended definition of VotingShareholder.

In addition, labels have been added to most of the properties and a few classes in the ontology, and several literals were revised to add the appropriate datatypes, primarily xsd:string and xsd:anyURI, depending on the annotation value.

Finally, the VOM ancillary model, from which the RDF/XML serialized OWL and UML XMI are generated, was revised to correct the duplicate proxy class for IncorporatedCompany.