BE-13, FBEF-38 - Summary: Review of any state web site for defining a business call this "Corporation". Not all corporations are companies, however. The actual restrictions on this class are exactly those appropriate for Corporation, but the text definition should be something like:
"A corporation is a separate legal entity owned by shareholders who enjoy protection from personal liability. Corporations are taxed annually on their earnings; corporate shareholders pay individual income tax on these earnings when they are distributed as dividends.
A corporation is managed by or under the direction of a board of directors, which generally determines corporate policy. Officers manage the day-to-day affairs of the corporation. Shareholders do not participate in day-to-day management activities. Management structure can be altered by committees of board members and shareholder agreements. Shareholders generally are not personally liable for obligations of the corporation."

Definition origin: http://www.business.ca.gov/StartaBusiness/DefiningaBusiness/Corporation.aspx

The Corporation class should have the following subclasses:
General Stock Corporation
Non-Profit Corporation
Benefit Corporation (which we have already in Functional Entities)
Common Interest Development Corporation

Non-Profit / benefit corporations have a corporate structure that can be defined using restrictions, and are not functional entities. They should be defined as kinds of corporations.

Discussion:
This issue affects section 9.2.3, Corporate Bodies, section 9.3.1, Corporations Ontology, in the FIBO BE 1.0 Alpha specification, with downstream impact on section 9.7.1, Functional Entities, section 9.6.3, Corporate Ownership, section 9.6.4, Corporate Control, and section 9.4.1, Partnerships.

Further review of the current set of concepts and related definitions has led to the consensus that the BodyCorporate concept in the Corporate Bodies ontology and the concept of a Corporation, at the highest level of abstraction, are synonymous.  Thus, the consensus is to rename BodyCorporate to Corporation, and integrate a source-able definition for it from a respected legal source, as long as that definition is essentially the same as what we currently have and is not jurisdiction-specific.  Links to jurisdiction-specific definitions may be provided, however.

Additionally, a better name for the current IncorporatedCompany concept is GeneralStockCorporation, which corresponds to the broader meaning of the semantics in the current ontology for incorporated company.  Again a better terminological definition should be sourced from an appropriate reference, as long as it is in keeping with the current text definition.

*** The resolution of this issue assumes that the resolution to issues FBEF-3, resolving mechanical syntax issues, issues FBEF-4, FBEF-6, FBEF-7, and FBEF-9 through FBEF-12, which resolve issues related to modifications in FIBO FND by the 1.0 finalization process, as well as issues FBEF-20 through FBEF-24, FBEF-30, FBEF-31, FBEF-34, and FBEF-35, which resolve lint errors, reasoning issues due to the hasRole property, and miscellaneous annotation issues have already been applied.