BE-117 -- modifications to the preliminary version of GovernmentEntities include: 1. Polity -- Changed parent from JuridicalPerson to LegalPerson -- not all polities are "legal entities" if the definition is that they have legal entity identifiers per ISO 17442, especially in the case of polities for small local governments in some parts of the world
2.  GovernmentalEntity --
changed name to GovernmentBody -- not all government bodies are "entities" -- especially if one considers departments with governments to be government bodies
changed parent to FormalOrganization (again because not all government bodies are legal entities in their own right)

3.  LegislativeBranch --
changed name to Legislature, which corresponds to the definition and use of the term internationally

4.  JudicialBranch --
changed name to Judiciary -- for the same reason as the change for Legislature

5.  RegionalState --
changed name to RegionalSovereignty -- to limit confusion with FederalState, etc.
changed the restriction with respect to shared sovereignty to say that it has shared sovereignty over some GeopoliticalArea, corresponding to the definition, which maybe broader than that of a federal state

6.  Eliminated children of Instrumentality -- these were redundant.  We have the notion of a chartered legal person (or body), which is a legal person chartered by the crown -- some instrumentalities will be both instrumentalities and chartered legal persons
We have the notion of a statutory body (and can add statutory corporation if that's really needed, but probably not necessary) -- some instrumentalities will be statutory bodies and instrumentalities, and some will be statutory bodies, corporations, and instrumentalities, and some will be statutory bodies and corporations but not instrumentalities, such as the Corporation for Public Broadcasting

Note that the definition of Instrumentality has been revised to say that an instrumentality is a government body ..., to make it clear that the concept of an instrumentality only applies to government, not to corporations that may have been created through legislation and may derive their creditworthiness from the government but are independent, such as Fannie Mae and Freddie Mac.  Fannie Mae and Freddie Mac are statutory bodies and stock corporations, also known as government-sponsored enterprises (GSEs) in the US, but are NOT instrumentalities by definition.

Government chartered and owned corporations, on the other hand, may be instrumentalities, such as the Federal Deposit Insurance Corporation (FDIC).

Individuals can be multiply classified using combinations of these classes, as needed.